% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_format.R
\name{d3_format}
\alias{d3_format}
\title{d3.format}
\usage{
d3_format(specifier)
}
\arguments{
\item{specifier}{A specifier to format value(s),
see \url{https://github.com/d3/d3-format#format} for details.}
}
\value{
a character vector
}
\description{
Format numbers for human consumption
 with \href{https://github.com/d3/d3-format}{d3.format}.
}
\examples{

x <- c(123456.789, 987654.321)
d3_format(",.2s")(x)


# Percentage
d3_format(".0\%")(0.123)
d3_format(".1\%")(0.123)
d3_format(".2\%")(0.123)


# Fixed width
d3_format("20")(c(1, 10, 100))

# Grouped thousands with two significant digits
d3_format(",.2r")(c(4223, 8445, 24556))

}
