% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{d3_format_js}
\alias{d3_format_js}
\title{Format numbers in htmlwidgets}
\usage{
d3_format_js(specifier, prefix = "", suffix = "", locale = "en-US")
}
\arguments{
\item{specifier}{Format for numbers, currency, percentage, e.g. \code{".0\%"} for rounded percentage.
See online documentation : \url{https://github.com/d3/d3-format}.}

\item{prefix}{Character string to append before formatted value.}

\item{suffix}{Character string to append after formatted value.}

\item{locale}{Localization to use, for example \code{"fr-FR"} for french,
see possible values here: \url{https://github.com/d3/d3-format/tree/master/locale}.}
}
\value{
a \code{JS} function
}
\description{
Format numbers in htmlwidgets
}
