% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{html_dependency_d3format}
\alias{html_dependency_d3format}
\title{HTML dependency for d3.format}
\usage{
html_dependency_d3format()
}
\value{
an \code{\link[htmltools]{htmlDependency}} object.
}
\description{
This can be used to add a dependency in \code{htmlwidget}
 but also to load d3.format in shiny applications or RMarkdown
}
\examples{
# This will return an HTML dependency
html_dependency_d3format()
# Using it in shiny or markdown will allow
# to use d3.format JavaScript API

\dontrun{

# when autoring htmlwidgets, you can
# add the dependency in createWidget:

createWidget(
  name = "mywidget",
  x = x,
  dependencies = list(
    html_dependency_d3format()
  )
)

}

}
