% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{use_d3_format}
\alias{use_d3_format}
\title{Use d3.format with a htmlwidget}
\usage{
use_d3_format(widget)
}
\arguments{
\item{widget}{An object of type \code{htmlwidget}.}
}
\value{
The same \code{htmlwidget} with an extra dependency.
}
\description{
Use d3.format with a htmlwidget
}
